Update PelicanFS dependency, add unit tests, futz with actions
I mistakenly pushed/published a version of this plugin that used a
version of PelicanFS that does not work for all use cases.

Specifically, the plugin would have failed whenever PelicanFS was
used to check the existence of a non-existent object.

A new PelicanFS release was cut with my bugfix, so this change indicates
we should use it.

Additionally, I used some agentic AI to add a set of unit tests that
handle PelicanFS mocking. I also disabled an action that imposed
restrictions on how I title my PRs -- I don't like that this was
built into the poetry scaffolding.